% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalBias.R
\name{cond_bias}
\alias{cond_bias}
\title{Compute conditional bias}
\usage{
cond_bias(y, pk, pkl, n, N, sample = TRUE, sampling = c("ups", "srs",
  "poisson"))
}
\arguments{
\item{y}{vector of observations of the variable of interest}

\item{pk}{vector of first-order inclusion probabilities; used only with
\code{sampling="ups"} and \code{sampling="poisson"}}

\item{pkl}{matrix of joint-inclusion probabilities; used only with \code{sampling="ups"}}

\item{n}{integer, the sample size; used only with \code{sampling="srs"}}

\item{N}{integer, the population size; used only with \code{sampling="srs"}}

\item{sample}{boolean, should the sample estimator be computed? Ignored for \code{sampling="poisson"}}

\item{sampling}{string indicating whether the conditional bias is to be computed
for a generic unequal probability sampling design (\code{"ups"}), a simple
random sampling (\code{"srs"}) or for Poisson sampling (\code{"poisson"}).}
}
\value{
a numeric vector of conditional bias
}
\description{
Compute bias of the Horvitz-Thompson estimator conditionally to sample units
}
