% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_cmin.R
\name{Delta_cmin}
\alias{Delta_cmin}
\title{Compute Delta(cmin)}
\usage{
Delta_cmin(cBias)
}
\arguments{
\item{cBias}{vector of conditional bias, computed by function \code{conditional_bias()}}
}
\value{
a scalar, representing the Delta value corresponding to the minimum c
}
\description{
Compute Delta(cmin) quantity for the Robust Horvitz-Thompson estimator
}
\examples{
# Generate population data
N <- 50; n <- 5

set.seed(0)
x <- rgamma(500, scale=10, shape=5)
y <- abs( 2*x + 3.7*sqrt(x) * rnorm(N) )

# Select sample
pik <- n * x/sum(x)
s   <- sample(N, n)
ys <- y[s]
piks <- pik[s]

# Compute conditional bias
cb <- conditional_bias(y=ys, pk=piks, sampling = "poisson")

# Find Delta(cmin)
Delta_cmin(cb)

}
