% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_c.R
\name{find_cmin_iter}
\alias{find_cmin_iter}
\title{Find the minimum c value iteratively}
\usage{
find_cmin_iter(cBias, niter = 1000)
}
\arguments{
\item{cBias}{vector of conditional bias for a sample}
}
\value{
a scalar, representing the value c that minimizes the maximum Bias
    of the robust HT estimator over a grid of values.
}
\description{
Compute the value c that minimizes the maximum absolute conditional Bias of
the robust HT estimator over a grid of values.
}
