% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHTestimator.R
\name{RHTestimate}
\alias{RHTestimate}
\title{Robust Horvitz-Thompson estimator}
\usage{
RHTestimate(ys, pks, cBias, cmin, Delta_cmin)
}
\arguments{
\item{ys}{vector of sample observed values}

\item{pks}{vector of sample inclusion probabilities}

\item{cBias}{vector of conditional bias, used if \code{Delta_cmin} is missing}

\item{Delta_cmin}{Delta value, corresponding to the minimum value of the tuning
constant c, computed by function \code{min_Delta()}. Either this or the combination
of \code{cBiad} and \code{c} should be used.}

\item{c}{tuning parameter to compute robust estimate, used if Delta_cmin is missing}
}
\value{
The Robust Horvitz-Thompson estimate
}
\description{
Estimate the Robust Horvitz-Thompson estimator according to
Beaumont, Haziza, Ruiz-Gazen (2013)
}
