% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHTestimator.R
\name{Delta_cmin}
\alias{Delta_cmin}
\title{Compute Delta(cmin) quantity for the Robust Horvitz-Thompson estimator}
\usage{
Delta_cmin(cBias)
}
\arguments{
\item{cBias}{vector of conditional bias, computed by function \code{conditional_bias()}}
}
\value{
a scalar, representing the Delta value corresponding to the minimum c
}
\description{
Compute Delta(cmin) quantity for the Robust Horvitz-Thompson estimator
}
