% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_c.R
\name{find_cmin}
\alias{find_cmin}
\title{Find the minimum c value over a grid of values}
\usage{
find_cmin(cBias, ngrid = 4 * length(cBias))
}
\arguments{
\item{cBias}{vector of conditional bias for a sample}

\item{ngrid}{integer scalar indicating the length of the grid of values that
is generated to estimate the optimum value for c}
}
\value{
A scalar, representing the value c that minimizes the maximum Bias
    of the robust HT estimator over a grid of values.
}
\description{
Compute the value c that minimizes the maximum absolute conditional Bias of
the robust HT estimator over a grid of values.
}
\examples{
# Generate population data
N <- 50; n <- 5

set.seed(0)
x <- rgamma(500, scale=10, shape=5)
y <- abs( 2*x + 3.7*sqrt(x) * rnorm(N) )

# Select sample
pik <- n * x/sum(x)
s   <- sample(N, n)
ys <- y[s]
piks <- pik[s]

# Compute conditional bias
cb <- conditional_bias(y=ys, pk=piks, sampling = "poisson")

# Find the minimum c
find_cmin(cb, ngrid = 200)


}
